changed tests in folders
