+++ cypress/e2e/settings/users/fees&fines/manual-charges.cy.js
@@ -1,0 +1,67 @@
+import { NOTICE_CATEGORIES } from '../../../../support/fragments/settings/circulation/patron-notices/noticePolicies';
+import NoticeTemplates from '../../../../support/fragments/settings/circulation/patron-notices/noticeTemplates';
+import ManualCharges from '../../../../support/fragments/settings/users/manualCharges';
+import UsersOwners from '../../../../support/fragments/settings/users/usersOwners';
+import SettingsMenu from '../../../../support/fragments/settingsMenu';
+
+describe('Fees&Fines', () => {
+  describe('Settings Users (Fee/fine)', () => {
+    const owners = [UsersOwners.getDefaultNewOwner(), UsersOwners.getDefaultNewOwner()];
+    const templates = [
+      NoticeTemplates.getDefaultTemplate({ category: NOTICE_CATEGORIES.FeeFineCharge }),
+      NoticeTemplates.getDefaultTemplate({ category: NOTICE_CATEGORIES.FeeFineAction }),
+    ];
+    const manualCharge = ManualCharges.getManualCharge({
+      owner: owners[0],
+      chargeNoticeId: templates[0].name,
+      actionNoticeId: templates[1].name,
+    });
+
+    before('Create test data', () => {
+      cy.getAdminToken().then(() => {
+        owners.forEach((owner) => UsersOwners.createViaApi(owner));
+        templates.forEach((template) => NoticeTemplates.createViaApi(template));
+      });
+
+      cy.loginAsAdmin({
+        path: SettingsMenu.manualCharges,
+        waiter: ManualCharges.waitLoading,
+      });
+    });
+
+    after('Delete test data', () => {
+      templates.forEach(({ id }) => {
+        NoticeTemplates.deleteViaApi(id);
+      });
+      owners.forEach(({ id }) => {
+        UsersOwners.deleteViaApi(id);
+      });
+    });
+
+    it(
+      'C442 Verify that you can create/edit/delete manual charges for a fee/fine owner (vega) (TaaS)',
+      { tags: ['extendedPath', 'vega'] },
+      () => {
+        ManualCharges.selectOwner(owners[0]);
+        ManualCharges.createViaUi(manualCharge);
+        manualCharge.amount = '$10.00';
+        ManualCharges.checkResultsTableContent([manualCharge]);
+
+        ManualCharges.createViaUi(manualCharge);
+        ManualCharges.checkValidatorError({ error: 'Fee/fine type already exists' });
+        ManualCharges.clickCancelBtn();
+
+        const editedManualCharge = {
+          ...manualCharge,
+          feeFineType: `Edited: ${manualCharge.feeFineType}`,
+          amount: '100.00',
+        };
+        ManualCharges.editViaUi(manualCharge, editedManualCharge);
+        editedManualCharge.amount = '$100.00';
+        ManualCharges.checkResultsTableContent([editedManualCharge]);
+
+        ManualCharges.deleteViaUi(editedManualCharge);
+      },
+    );
+  });
+});
