+++ cypress/e2e/settings/users/fees&fines/owners.cy.js
@@ -1,9 +1,9 @@
-import Permissions from '../../../support/dictionary/permissions';
-import ServicePoints from '../../../support/fragments/settings/tenant/servicePoints/servicePoints';
-import UsersOwners from '../../../support/fragments/settings/users/usersOwners';
-import SettingsMenu from '../../../support/fragments/settingsMenu';
-import users from '../../../support/fragments/users/users';
-import getRandomPostfix from '../../../support/utils/stringTools';
+import Permissions from '../../../../support/dictionary/permissions';
+import ServicePoints from '../../../../support/fragments/settings/tenant/servicePoints/servicePoints';
+import UsersOwners from '../../../../support/fragments/settings/users/usersOwners';
+import SettingsMenu from '../../../../support/fragments/settingsMenu';
+import users from '../../../../support/fragments/users/users';
+import getRandomPostfix from '../../../../support/utils/stringTools';
 
 // TO DO: remove ignoring errors. Now when you click on one of the buttons, some promise in the application returns false
 Cypress.on('uncaught:exception', () => false);
